2009-05-13  Luis Fernando Muñoz Mejías
	* Release: 1.5.0
	- Add a hack to control which files should be managed by the
	component.

	CERN was using different tools for handling the
	.ssh/authorized_keys file, and was colliding with this
	component. The proper way to solve it would be by disabling
	this component on the affected users, but they needed it to
	control PAM ACLs. And PAM ACLs should be moved to a different
	component but that needs some time I don't have, so this is
	the quick hack.



git-svn-id: https://quattor.svn.sourceforge.net/svnroot/quattor/trunk/ncm-components/core/ncm-useraccess@906 b8b08b4b-6333-482b-9f54-fdf6c5fb3d79

+++ ncm-useraccess/src/main/perl/useraccess.pm
@@ -5,7 +5,7 @@
 # File: useraccess.pm
 # Implementation of ncm-useraccess
 # Author: Luis Fernando Muñoz Mejías <mejias@delta.ft.uam.es>
-# Version: 1.4.6 : 10/02/09 11:39
+# Version: 1.5.0 : 13/05/09 18:21
 # 
 #
 # Note: all methods in this component are called in a
@@ -55,6 +55,9 @@ use constant REALM	=> "realm";
 use constant PRINCIPAL=> "principal";
 use constant INSTANCE	=> "instance";
 use constant HOST	=> "host";
+
+# Field containing the credentials the component should manage:
+use constant MANAGED_CREDENTIALS => "managed_credentials";
 
 # getpwnams values.
 use constant NAME	=> 0;
@@ -316,18 +319,32 @@ sub files
     $h{SSH_KEYS()} = CAF::FileWriter->new($path, log => $self,
 					  owner => $uid,
 					  group => $gid);
+
+    foreach my $cred ($uconfig->{MANAGED_CREDENTIALS()}) {
+	$h{MANAGED_CREDENTIALS()}->{$cred} = 1;
+    }
+
     return \%h;
 }
 
-# Closes the opened files.
+# Closes the opened files. If they are expected to be handled by the
+# component this includes saving the contents or removing empty
+# files. Otherwise, the contents are just cancelled.
 sub close_files
 {
     my ($self, $f) = @_;
 
-    foreach my $fh (values (%$f)) {
-	my $cnt = $fh->string_ref();
-	unless ($$cnt) {
-	    unlink(*$fh->{filename});
+    my $mg = $f->{MANAGED_CREDENTIALS()};
+    delete($f->{MANAGED_CREDENTIALS()});
+
+    while (my ($k, $fh) = each(%$f)) {
+	if ($mg->{$k}) {
+	    my $cnt = $fh->string_ref();
+	    unless ($$cnt) {
+		unlink(*$fh->{filename});
+		$fh->cancel();
+	    }
+	} else {
 	    $fh->cancel();
 	}
 	$fh->close();
@@ -357,9 +374,9 @@ sub Configure
 	my $fhash = $self->files($uconfig, $uid, $gid, $home);
 	if ($self->set_kerberos($user, $uconfig, $fhash) ||
 	    $self->set_ssh_fromurls($user, $uconfig,
-				    $fhash->{SSH_KEYS()}) ||
+	    $fhash->{SSH_KEYS()}) ||
 	    $self->set_ssh_fromkeys($user, $uconfig,
-				    $fhash->{SSH_KEYS()}) ||
+	    $fhash->{SSH_KEYS()}) ||
 	    $self->set_acls($user, $uconfig) ||
 	    $self->set_roles($user, $uconfig->{ROLES()},
 			     $rlhash, $fhash)) {
